Upadated handlebars and routes.
