+++ controllers/ultraportablesRoutes.js
@@ -1,0 +1,2 @@
+const router = require('express').Router();
+module.exports = router;
