+++ views/home.handlebars
@@ -1,7 +1,7 @@
   <header>
         <nav class="navbar navbar-expand-lg navbar-dark fixed-top bg-dark bg-primary">
             <div class="container">
-                <a class="navbar-brand" href="/home">
+                <a class="navbar-brand" href="home">
                     <img src="../../images/icons8-shopping-bag-50.png" width="30" height="30" class="d-inline-block align-top"
                         alt="VERPUR Shop Logo">
                     VERPUR Shop
@@ -15,7 +15,7 @@
                 <div class="collapse navbar-collapse" id="navbarCollapse">
                     <ul class="nav navbar-nav mr-auto">
                         <li class="nav-item ">
-                            <a class="nav-link active" href="/home">Home</a>
+                            <a class="nav-link active" href="home">Home</a>
                         </li>
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle" href="#" id="nav-dropdown" data-bs-toggle="dropdown"
@@ -23,28 +23,28 @@
                                 Laptops
                             </a>
                             <ul class="dropdown-menu" aria-labelledby="nav-dropdown">
-                                <li><a class="dropdown-item" href="/ultrabooks">Ultrabooks</a></li>
-                                <li><a class="dropdown-item" href="/ultraportables">Ultraportable</a></li>
-                                <li><a class="dropdown-item" href="/macbooks">MacBooks</a></li>
+                                <li><a class="dropdown-item" href="ultrabooks">Ultrabooks</a></li>
+                                <li><a class="dropdown-item" href="ultraportables">Ultraportable</a></li>
+                                <li><a class="dropdown-item" href="macbooks">MacBooks</a></li>
                                 <li>
                                     <hr class="dropdown-divider">
                                 </li>
-                                <li><a class="dropdown-item" href="/alllaptops">All Laptops</a></li>
+                                <li><a class="dropdown-item" href="alllaptops">All Laptops</a></li>
                             </ul>
                         </li>
                        
                         <li class="nav-item">
-                            <a class="nav-link " href="/subscription">
+                            <a class="nav-link " href="subscription">
                                 laptop subscription
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link " href="/login">
+                            <a class="nav-link " href="login">
                                 Login
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="/contact">Contact</a>
+                            <a class="nav-link" href="contact">Contact</a>
                         </li>
                     </ul>
                 </div>
@@ -74,7 +74,7 @@
                     <div class="carousel-caption text-start">
                         <h1>Discover our subscription plans</h1>
                         <p>Our Laptops delivered to your door.</p>
-                        <p><a class="btn btn-lg btn-primary" href="/subscription" role="button">Sign up
+                        <p><a class="btn btn-lg btn-primary" href="subscription" role="button">Sign up
                                 today</a></p>
                     </div>
                 </div>
@@ -82,7 +82,7 @@
                     <div class="carousel-caption text-end">
                         <h1>Browse laptops</h1>
                         <p>Every person needs a laptop. </p>
-                        <p><a class="btn btn-lg btn-primary" href="/alllaptops" role="button">Laptops</a></p>
+                        <p><a class="btn btn-lg btn-primary" href="alllaptops" role="button">Laptops</a></p>
                     </div>
                 </div>
             </div>
@@ -115,7 +115,7 @@
                             <p class="card-text">Our famous ultrabooks!</p>
                         </div>
                         <div class="card-footer">
-                            <a href="/product-details" class="btn btn-primary">View details</a>
+                            <a href="product-details" class="btn btn-primary">View details</a>
                         </div>
                     </div>
                 </div>
@@ -128,7 +128,7 @@
                             <p class="card-text">My God, so Ultra!</p>
                         </div>
                         <div class="card-footer">
-                            <a href="/product-details" class="btn btn-primary">View details</a>
+                            <a href="product-details" class="btn btn-primary">View details</a>
                         </div>
                     </div>
                 </div>
@@ -141,7 +141,7 @@
                             <p class="card-text">Our populer MacBooks!</p>
                         </div>
                         <div class="card-footer">
-                            <a href="/product-details" class="btn btn-primary">View details</a>
+                            <a href="product-details" class="btn btn-primary">View details</a>
                         </div>
                     </div>
                 </div>
