+++ views/layouts/main.handlebars
@@ -9,11 +9,9 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
 
-    <link href="../../public/css/bootstrap-override.css" rel="stylesheet">
-
+    <link rel="stylesheet" href="../../public/css/bootstrap-override.css" >
     <link rel="stylesheet" href="../../public/css/custom.css">
     <link rel="stylesheet" href="../../public/css/sign-in.css">
-    <link rel="stylesheet" href="../../public/css/bootstrap-override.css">
 </head>
 
 <body>
