+++ views/macbooks.handlebars
@@ -1,7 +1,7 @@
  <header>
         <nav class="navbar navbar-expand-lg navbar-dark fixed-top bg-dark bg-primary">
             <div class="container">
-                    <a class="navbar-brand" href="/home">
+                    <a class="navbar-brand" href="home">
                     <img src="../../images/icons8-shopping-bag-50.png" width="30" height="30" class="d-inline-block align-top"
                         alt="VERPUR Shop Logo">
                     VERPUR Shop
@@ -15,7 +15,7 @@
                 <div class="collapse navbar-collapse" id="navbarCollapse">
                     <ul class="nav navbar-nav mr-auto">
                         <li class="nav-item">
-                            <a class="nav-link" href="/home">Home</a>
+                            <a class="nav-link" href="home">Home</a>
                         </li>
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle" href="#" id="nav-dropdown" data-bs-toggle="dropdown"
@@ -23,22 +23,22 @@
                                 Pies
                             </a>
                             <ul class="dropdown-menu" aria-labelledby="nav-dropdown">
-                                 <li><a class="dropdown-item" href="/ultrabooks">Ultrabooks</a></li>
-                                <li><a class="dropdown-item" href="/ultraportables">Ultraportable</a></li>
-                                <li><a class="dropdown-item" href="/macbooks">MacBooks</a></li>
+                                 <li><a class="dropdown-item" href="ultrabooks">Ultrabooks</a></li>
+                                <li><a class="dropdown-item" href="ultraportables">Ultraportable</a></li>
+                                <li><a class="dropdown-item" href="macbooks">MacBooks</a></li>
                                 <li>
                                     <hr class="dropdown-divider">
                                 </li>
-                                <li><a class="dropdown-item" href="/alllaptops">All pies</a></li>
+                                <li><a class="dropdown-item" href="alllaptops">All pies</a></li>
                             </ul>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="/subscription">
+                            <a class="nav-link" href="subscription">
                                 Pie subscription
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="/contact">Contact</a>
+                            <a class="nav-link" href="contact">Contact</a>
                         </li>
                     </ul>
                 </div>
@@ -55,7 +55,7 @@
         <div class="container">
             <nav class="my-3 ms-3" style="--bs-breadcrumb-divider: '>';">
                 <ol class="breadcrumb">
-                    <li class="breadcrumb-item"><a href="/">Home</a></li>
+                    <li class="breadcrumb-item"><a href="home">Home</a></li>
                     <li class="breadcrumb-item"><a href="#">Laptops</a></li>
                     <li class="breadcrumb-item active" aria-current="page">MacBooks </li>
                 </ol>
@@ -69,7 +69,7 @@
                             <p class="card-text">You'll love it! </p>
                         </div>
                         <div class="card-footer">
-                            <a href="/product-details" class="btn btn-primary">View details</a>
+                            <a href="product-details" class="btn btn-primary">View details</a>
                         </div>
 
                     </div>
@@ -83,7 +83,7 @@
                             <p class="card-text">Plain cheese cake. Plain pleasure.</p>
                         </div>
                         <div class="card-footer">
-                            <a href="/product-details" class="btn btn-primary">View details</a>
+                            <a href="product-details" class="btn btn-primary">View details</a>
                         </div>
                     </div>
                 </div>
@@ -96,7 +96,7 @@
                             <p class="card-text">You'll love it! </p>
                         </div>
                         <div class="card-footer">
-                            <a href="/product-details" class="btn btn-primary">View details</a>
+                            <a href="product-details" class="btn btn-primary">View details</a>
                         </div>
                     </div>
                 </div>
