+++ views/product-details.handlebars
@@ -1,7 +1,7 @@
   <header>
         <nav class="navbar navbar-expand-lg navbar-dark fixed-top bg-dark bg-primary">
             <div class="container">
-                    <a class="navbar-brand" href="/home">
+                    <a class="navbar-brand" href="home">
                     <img src="../../images/icons8-shopping-bag-50.png" width="30" height="30" class="d-inline-block align-top"
                         alt="VERPUR Shop Logo">
                     VERPUR Shop
@@ -15,7 +15,7 @@
                 <div class="collapse navbar-collapse" id="navbarCollapse">
                     <ul class="nav navbar-nav mr-auto">
                         <li class="nav-item">
-                            <a class="nav-link" href="/home">Home</a>
+                            <a class="nav-link" href="home">Home</a>
                         </li>
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle" href="#" id="nav-dropdown" data-bs-toggle="dropdown"
@@ -23,27 +23,27 @@
                                 Pies
                             </a>
                             <ul class="dropdown-menu" aria-labelledby="nav-dropdown">
-                             <li><a class="dropdown-item" href="/ultrabooks">Ultrabooks</a></li>
-                                <li><a class="dropdown-item" href="/ultraportables">Ultraportable</a></li>
-                                <li><a class="dropdown-item" href="/macbooks">MacBooks</a></li>
+                             <li><a class="dropdown-item" href="ultrabooks">Ultrabooks</a></li>
+                                <li><a class="dropdown-item" href="ultraportables">Ultraportable</a></li>
+                                <li><a class="dropdown-item" href="macbooks">MacBooks</a></li>
                                 <li>
                                     <hr class="dropdown-divider">
                                 </li>
-                                <li><a class="dropdown-item" href="/alllaptops">All pies</a></li>
+                                <li><a class="dropdown-item" href="alllaptops">All pies</a></li>
                             </ul>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="/subscription">
+                            <a class="nav-link" href="subscription">
                                 Pie subscription
                             </a>
                         </li>
                           <li class="nav-item">
-                            <a class="nav-link " href="/login">
+                            <a class="nav-link " href="login">
                                 Login
                             </a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link" href="/contact">Contact</a>
+                            <a class="nav-link" href="contact">Contact</a>
                         </li>
                     </ul>
                 </div>
@@ -62,9 +62,9 @@
 
             <nav class="my-3 ms-3" style="--bs-breadcrumb-divider: '>';">
                 <ol class="breadcrumb">
-                    <li class="breadcrumb-item"><a href="/">Home</a></li>
+                    <li class="breadcrumb-item"><a href="home">Home</a></li>
                     <li class="breadcrumb-item"><a href="#">Laptops</a></li>
-                    <li class="breadcrumb-item"><a href="Ultrabooks">Ultrabooks</a></li>
+                    <li class="breadcrumb-item"><a href="ultrabooks">Ultrabooks</a></li>
                     <li class="breadcrumb-item active" aria-current="page">Ultrabook</li>
                 </ol>
             </nav>
@@ -285,7 +285,7 @@
                         </div>
                     </div>
                 </form>
-                <a type="button" href="order.html" class="m-1 btn btn-success">
+                <a type="button" href="order" class="m-1 btn btn-success">
                     Go to order page
                 </a>
