+++ views/signup.handlebars
@@ -1,25 +1,69 @@
-<main class="form-signin w-100 m-auto">
-  <form>
-    <img class="mb-4" src="../images/icons8-purple-shopping-bag-50.png" alt="" width="72" height="57">
-    <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
+  <img class="mb-4" src="../images/icons8-purple-shopping-bag-50.png" alt="" width="72" height="57">
+    <h1 class="h3 mb-3 fw-normal">Please sign up</h1>
 
-    <div class="form-floating">
-      <input type="email" class="form-control" id="floatingInput" placeholder="name@example.com">
-      <label for="floatingInput">Email address</label>
-    </div>
-    <div class="form-floating">
-      <input type="password" class="form-control" id="floatingPassword" placeholder="Password">
-      <label for="floatingPassword">Password</label>
-    </div>
+<form class="row g-3">
 
-    <div class="checkbox mb-3">
-      <label>
-        <input type="checkbox" value="remember-me"> Remember me
+ <div class="col-md-6">
+    <label for="inputFirstname4" class="form-label">First Name</label>
+    <input type="text" class="form-control" id="inputFirstname4">
+  </div>
+   <div class="col-md-6">
+    <label for="inputLastname4" class="form-label">Last Name</label>
+    <input type="text" class="form-control" id="inputLastname4">
+  </div>
+ <div class="col-md-6">
+    <label for="inputUsername4" class="form-label">Username</label>
+    <input type="email" class="form-control" id="inputUsername4">
+  </div>
+   <div class="col-md-6">
+    <label for="inputPassword4" class="form-label">Password</label>
+    <input type="password" class="form-control" id="inputPassword4">
+  </div>
+  <div class="col-md-6">
+    <label for="inputEmail4" class="form-label">Email</label>
+    <input type="email" class="form-control" id="inputEmail4">
+  </div>
+  <div class="col-md-6">
+    <label for="inputPhone4" class="form-label">Phone</label>
+    <input type="text" class="form-control" id="inputPhone4">
+  </div>
+  <div class="col-12">
+    <label for="inputAddress" class="form-label">Address</label>
+    <input type="text" class="form-control" id="inputAddress" placeholder="1234 Main St">
+  </div>
+  <div class="col-12">
+    <label for="inputAddress2" class="form-label">Address 2</label>
+    <input type="text" class="form-control" id="inputAddress2" placeholder="Apartment, studio, or floor">
+  </div>
+  <div class="col-md-6">
+    <label for="inputCity" class="form-label">City</label>
+    <input type="text" class="form-control" id="inputCity">
+  </div>
+  <div class="col-md-4">
+    <label for="inputState" class="form-label">State</label>
+    <select id="inputState" class="form-select">
+      <option selected>Choose...</option>
+      <option>...</option>
+    </select>
+  </div>
+  <div class="col-md-2">
+    <label for="inputZip" class="form-label">Zip</label>
+    <input type="text" class="form-control" id="inputZip">
+  </div>
+  <div class="col-md-2">
+    <label for="inputCountry" class="form-label">Country</label>
+    <input type="text" class="form-control" id="inputCountry">
+  </div>
+
+  <div class="col-12">
+    <div class="form-check">
+      <input class="form-check-input" type="checkbox" id="gridCheck">
+      <label class="form-check-label" for="gridCheck">
+        Select for Premium Membership
       </label>
     </div>
-    <button class="w-100 btn btn-lg btn-primary" type="submit">Sign in</button>
-    <p class="mt-5 mb-3 text-muted">&copy; 2017–2023</p>
-  </form>
-</main>
-
-<script src="/public/js/login.js"></script>
+  </div>
+  <div class="col-12">
+    <button type="submit" class="btn btn-primary">Sign in</button>
+  </div>
+</form>
